Add tests for field_map and field_spec filtering.
Fix bug that cropped up when using multiple filters in a call to field_map.

Made field_spec instances callable; pass in filters, and a new
FieldSpecMap is returned based on the original instance, but
containing only those keys and values where the value is included
by the filter(s).